Refactored ExplorerView
