+++ .gitignore
@@ -48,3 +48,4 @@ Docs/.Rhistory
 Documentation/errors.txt
 .vs/ApsimX/v15/Server/sqlite3/storage.ide-wal
 .vs/ApsimX/v15/Server/sqlite3/storage.ide-shm
+*.db
