+++ ApsimNG/ApsimNG.csproj
@@ -207,6 +207,9 @@
     <Compile Include="Interfaces\IButtonView.cs" />
     <Compile Include="Interfaces\ICommand.cs" />
     <Compile Include="Interfaces\ICultivarView.cs" />
+    <Compile Include="Interfaces\IMenuView.cs" />
+    <Compile Include="Interfaces\IToolStripView.cs" />
+    <Compile Include="Interfaces\ITreeView.cs" />
     <Compile Include="Interfaces\IExplorerView.cs" />
     <Compile Include="Interfaces\IGraphView.cs" />
     <Compile Include="Interfaces\IGridCell.cs" />
@@ -299,6 +302,9 @@
     <Compile Include="Views\EditBoxView.cs" />
     <Compile Include="Views\EditorView.cs" />
     <Compile Include="Views\ErrorView.cs" />
+    <Compile Include="Views\MenuView.cs" />
+    <Compile Include="Views\ToolStripView.cs" />
+    <Compile Include="Views\TreeView.cs" />
     <Compile Include="Views\ExplorerView.cs" />
     <Compile Include="Views\FactorControlView.cs" />
     <Compile Include="Views\FolderView.cs" />
